Support cmd-line overrides which include commas
While perfrunner allows users to override test settings at the command
line, it isn't possible to override settings which include a comma in
them - even if quotes or escapes are used - for example:

    foo.var.value,"cluster.jemalloc_conf.lg_tcache_max:17,lg_dirty_mult:2"

Perfrunner naively splits on all commas regardless of their context.

Change the parsing to use the csv module, which can handle quoing of
strings, allowing the above to work.